delete file

+++ src/test.py
@@ -17,10 +17,10 @@ for data in testDataset:
     lr = data[0].numpy()
     hr = data[1][0].numpy()
     print(data[2])
-    lr = np.transpose(lr, (1,2,0)).astype(np.uint8)
+    lr = np.transpose(lr, (1,2,0))
     hr = np.transpose(hr, (1,2,0))
 
-    lr = np.array(Image.fromarray(lr).resize(hr.shape[:2], Image.BICUBIC))
+    lr = np.array(Image.fromarray(lr.astype(np.uint8)).resize(hr.shape[:2], Image.BICUBIC))
     #lr = scipy.misc.imresize(lr, hr.shape, interp='bicubic')
 
     psnr = utils.calc_psnr(lr, hr, scale=0, rgb_range=255.)
